warning

+++ ivl/utility/colors_example.cpp
@@ -1,10 +1,10 @@
-#include <ivl/utility/colors>
 #include <ivl/format>
+#include <ivl/utility/colors>
 
 int main() {
   namespace term = ivl::terminal_graphical_rendition;
   for (int i = 0; i < 256; i += 5) {
-    for (int j = 0; j < 256; j += 2) ivl::fmt::print("{} ", term::background_color{i, i, j});
+    for (int j = 0; j < 256; j += 2) ivl::fmt::print("{} ", term::background_color{(uint8_t)i, (uint8_t)i, (uint8_t)j});
     ivl::fmt::println("{}", term::background_reset{});
   }
   char foo[] = "hello\nworld";
